Runs IntelliJ code cleanup
- Remove redundant toString() calls
- Remove useless semi-colons
- Make attributes final
- Remove duplicate blank lines
- Remove redundant casts

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/resources/DatacenterStorage.java
@@ -162,7 +162,7 @@ public class DatacenterStorage {
         }
 
         for (final SanStorage storage : getStorageList()) {
-            if (storage.isAmountAvailable((long) file.getSize())) {
+            if (storage.isAmountAvailable(file.getSize())) {
                 storage.addFile(file);
                 return DataCloudTags.FILE_ADD_SUCCESSFUL;
             }
