+++ cloudsim-plus/src/main/java/org/cloudsimplus/faultinjection/HostFaultInjection.java
@@ -671,7 +671,7 @@ public class HostFaultInjection extends CloudSimEntity {
      * @return
      */
     public long getNumberOfFaults() {
-        return (long) faultsOfAllVmsByBroker.values().size();
+        return faultsOfAllVmsByBroker.values().size();
     }
 
     /**
